refactor(test): move SHEETS tests to tests/ subdirectory pattern
Converts info.rs → info/mod.rs and adds proper tests/mod.rs,
tests/success.rs, tests/failure.rs following the project test layout.
